Update String Split Method and Improve Parsing Functions
Refactored the string split method in utils.h to accept a template type for substrings. This has increased the versatility of the function allowing it to return a vector of substrings as either string objects or string_view objects. Additionally, in the TryParse function, introduced an optional construct to handle failed conversions more gracefully. A successful conversion now returns the parsed value wrapped in an optional object, while a failed conversion returns an empty optional object. An alternative Parse function has also been added to throw an exception when the conversion fails.