Refactor Mesh and File Scripts
The code for Mesh and File scripts have been cleaned up and refactored for improved readability and efficiency. The changes include the use of GL* types for uniformity, adjustment in precision of glm::vec from lowp to mediump, and simpler assertions in the File script. Code repetition was minimized and more efficient vector management techniques were used.